fastapi/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
dataclasses-json==0.5.7
decorator==5.1.1
exceptiongroup==1.1.1
fastapi==0.96.0
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
langchain==0.0.189
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.7
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.2
py==1.11.0
pyarrow==12.0.0
pycparser==2.21
pydantic==1.10.8
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
ratelimiter==1.2.0.post0
regex==2023.6.3
requests==2.28.2
retry==0.9.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.6
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
watchfiles==0.19.0
weaviate-client==3.19.2
websockets==11.0.3
yarl==1.9.2
